split chats code
